fix(gm): finalize-gm completeness gate, resonance catch, dashboard counts (#3268)
